+++ domains/med/src/main/kotlin/de/jfachwert/med/Hilfsmittelnummer.kt
@@ -59,7 +59,7 @@ open  class Hilfsmittelnummer
      * @return true oder false
      */
     override fun isValid(): Boolean {
-        return VALIDATOR.isVald(code)
+        return VALIDATOR.isValid(code)
     }
 
     /**