+++ domains/med/src/main/kotlin/de/jfachwert/med/PZN.kt
@@ -54,7 +54,7 @@ open class PZN
      * @return true oder false
      */
     override fun isValid(): Boolean {
-        return VALIDATOR.isVald(code)
+        return VALIDATOR.isValid(code)
     }
 
     /**