+++ domains/med/src/main/kotlin/de/jfachwert/med/SNOMED.kt
@@ -58,7 +58,7 @@ open class SNOMED
      * @return true oder false
      */
     override fun isValid(): Boolean {
-        return VALIDATOR.isVald(code)
+        return VALIDATOR.isValid(code)
     }
 
     val display: String = display