+++ domains/med/src/main/kotlin/de/jfachwert/med/Versichertennummer.kt
@@ -57,7 +57,7 @@ open class Versichertennummer
      * @return true oder false
      */
     override fun isValid(): Boolean {
-        return VALIDATOR.isVald(code)
+        return VALIDATOR.isValid(code)
     }